WIP: Add support for building OS X disk image installers (drag and drop).
Most functionality is implemented; working disk images can be built.
Still needs .DS_Store synthesis and more complete implementation of
license file synthesis - i.e. need to have RezDoc read button title
localizations from the plist and emit these as a STR# resource. Also
need to implement a mechanism to use Qbs's translations as fallbacks
but let the user override them piece by piece. Need to emit errors if
a license translation is incomplete. Also minor, but need a way to copy
a directory's contents to the root of the staging directory (as opposed
to simply copying the directory itself as a subdirectory of the staging
directory).

Multi-language localizations are not possible.

Rich text formatting is not supported for Traditional Chinese and
Japanese (this limitation may be lifted in the future).

Task-number: QBS-416
Change-Id: I95431e21becc3bea075a1468d5271a41c38bbd04