Implement partial bootstrapping
This allows a minimal Qbs to be bootstrapped such that it can then build
itself. Supports Windows (both MSVC and MinGW), macOS, and any other
Unix-like platform, although only Linux was specifically tested.

Note that a functioning Qt installation is still required for now, but
it at least allows the qmake build to be deleted and thus we gain some
additional self-respect.

Task-number: QBS-68
Change-Id: I77f40e816c270e67d45e96e66dfd6cb9df2f7c0f