***Improve support for Windows SxS assembly manifests
There's no reason to limit manifest embedding solely to the manifest
generated by link.exe. Additionally, this makes it easier to add content
to the manifest by adding files with manifest fragments rather than
having to manually compose them and create Windows Resource files to
apply them.

On MSVC, this allows the manifest tool to merge multiple manifests into
the final assembly without having to manually create a .rc file or
perform the merging.

On MinGW, this allows a manifest to be merged into the final assembly
without having to manually create a .rc file. Unlike MSVC, manifest
merging is not yet supported because the toolchain does not provide this
capability. Qbs may begin to emulate this capability at some point, but
for now specifying multiple manifest files in a product with MinGW will
result in an error.

***needs tests, also testing is tough due to manifest extraction

Change-Id: I12e2db809c8b638168567f4342288eac568a57ca